Fix assembler crashing on parsing expressions due to a regression
Also revert to use IntArray everywhere in place of ByteArrays when dealing with binary data.
While it should be more appropriate using Byte types to work with binary data, unfortunately the way they work in Kotlin is very nontrivial and they cause more pain than gain.

+++ app/src/main/kotlin/com/flykespice/chip8ide/ui/viewmodel/ScaffoldViewModel.kt
@@ -21,7 +21,7 @@ class ScaffoldViewModel(private val chip8IdeManager: Chip8IdeManager) : ViewMode
         }
     }
 
-    fun importROM(data: ByteArray) {
+    fun importROM(data: IntArray) {
         viewModelScope.launch {
             chip8IdeManager.importROM(data)
         }
